requirements.txt
accelerate==0.12.0
torchvision
transformers>=4.21.0
ftfy
tensorboard
modelcards
twilio
python-dotenv
redis[hiredis]
